fixed after plugin move
