+++ src/lib/djerba/plugins/wgts/cnv/cnv_template.html
@@ -6,7 +6,7 @@
 <%
 import djerba.render.constants as constants
 from djerba.util.image_to_base64 import converter
-from djerba.cnv_tools.json_to_html import html_builder
+from djerba.plugins.wgts.cnv_tools.json_to_html import html_builder
 %>
