+++ src/lib/djerba/plugins/wgts/snv_indel/snv_indel_template.html
@@ -4,8 +4,8 @@
 
 <!-- IMPORTS -->
 <%
-  from djerba.snv_indel_tools.render import html_builder
-  import djerba.snv_indel_tools.constants as constants
+  from djerba.plugins.wgts.snv_indel_tools.render import html_builder
+  import djerba.plugins.wgts.snv_indel_tools.constants as constants
 %>
