+++ src/lib/djerba/plugins/wgts/snv_indel/test/snv_indel.ini
@@ -1,7 +1,7 @@
 [core]
 
 [wgts.snv_indel]
-maf_file = /.mounts/labs/CGI/scratch/fbeaudry/reporting/djerba_test_data_lfs/snv_indel-plugin/test.maf.gz
+maf_file = /.mounts/labs/CGI/scratch/fbeaudry/reporting/djerba_test_data_lfs/wgs-snv_indel-plugin/test.maf.gz
 cna_file = /.mounts/labs/CGI/scratch/fbeaudry/reporting/djerba_test_data_lfs/report_example/data_CNA.txt 
 oncotree_code=COAD
 tumour_id = OCT_011488-TS
