+++ src/lib/djerba/plugins/wgts/snv_indel_tools/preprocess.py
@@ -14,10 +14,10 @@ from djerba.sequenza import sequenza_reader
 from djerba.util.subprocess_runner import subprocess_runner
 from djerba.extract.oncokb.annotator import oncokb_annotator
 from shutil import copyfile
-import djerba.snv_indel_tools.constants as constants 
+import djerba.plugins.wgts.snv_indel_tools.constants as constants 
 from djerba.plugins.base import plugin_base
 import pandas as pd
-import djerba.snv_indel_tools.constants as sic
+import djerba.plugins.wgts.snv_indel_tools.constants as sic
 import djerba.render.constants as rc
 
 class preprocess(logger):
